add firebase, add default pfp and gc pic, add url for dev and prod websocket

+++ src/interceptor.ts
@@ -13,7 +13,8 @@ export function loggingInterceptor(req: HttpRequest<unknown>, next: HttpHandlerF
         catchError((error: HttpErrorResponse) => {
             // if we get a 401 with from an api other than check credentials, redirect to login. this assumes bc of an expired session. 
             // tampering with csrf token and refreshing page makes auth-guard fail and redirect to login page.  
-            if (error.status === 401 && !error.url?.endsWith("check-credentials")) {
+            // router.url !== home so not logged in user can actually see the home and not be rerouted to login form due to failed 401 attempt to retrieve personal user info
+            if (error.status === 401 && !error.url?.endsWith("check-credentials") && router.url !== "/home") {
                 localStorage.clear()
                 router.navigate(['/login'])
             }
